+++ 3-period/java-subway-path-precourse-main/src/main/java/subway/controller/SubwayController.java
@@ -6,14 +6,14 @@ import java.util.List;
 import subway.command.FunctionCommand;
 import subway.command.ProcessingState;
 import subway.command.RouteCriteriaCommand;
-import subway.domain.Line;
-import subway.domain.LineRepository;
+import subway.domain.line.Line;
+import subway.domain.line.LineRepository;
 import subway.domain.Order;
-import subway.domain.PathFinder;
-import subway.domain.Route;
-import subway.domain.RoutesRepository;
-import subway.domain.Station;
-import subway.domain.StationRepository;
+import subway.domain.route.PathFinder;
+import subway.domain.route.Route;
+import subway.domain.route.RoutesRepository;
+import subway.domain.station.Station;
+import subway.domain.station.StationRepository;
 import subway.exception.ErrorMessage;
 import subway.exception.ExceptionHandler;
 import subway.view.InputView;
