+++ 3-period/java-subway-path-precourse-main/src/main/java/subway/domain/Order.java
@@ -1,5 +1,6 @@
 package subway.domain;
 
+import subway.domain.station.Station;
 import subway.exception.CustomIllegalArgumentException;
 import subway.exception.ErrorMessage;
