+++ 3-period/java-subway-path-precourse-main/src/main/java/subway/domain/line/Line.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.line;
 
 public class Line {
