+++ 3-period/java-subway-path-precourse-main/src/main/java/subway/domain/station/StationRepository.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.station;
 
 import static subway.exception.ErrorMessage.INVALID_STATION_NAME;
