+++ 3-period/java-subway-path-precourse-main/src/test/java/subway/domain/PathFinderTest.java
@@ -6,6 +6,11 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import subway.domain.route.PathFinder;
+import subway.domain.route.Route;
+import subway.domain.route.RoutesRepository;
+import subway.domain.station.Station;
+import subway.domain.station.StationRepository;
 
 class PathFinderTest {
